Fix one cast, add one cast.

+++ libparse/clk_computime.c
@@ -122,7 +122,8 @@ inp_computime(
 {
 	unsigned int rtc;
 
-	parseprintf(DD_PARSE, ("inp_computime(0x%lx, 0x%x, ...)\n", (long)parseio, ch));
+	parseprintf(DD_PARSE, ("inp_computime(0x%lx, 0x%x, ...)\n",
+                    (long unsigned)parseio, (unsigned int)ch));
 
 	switch (ch)
 	{
